Moved various methods up into the superclass and eliminated two duplicated classes.
