+++ rifServices/src/main/java/rifServices/dataStorageLayer/ms/MSSQLInvestigationManager.java
@@ -19,6 +19,7 @@ import rifServices.businessConceptLayer.Geography;
 import rifServices.businessConceptLayer.HealthTheme;
 import rifServices.businessConceptLayer.Investigation;
 import rifServices.businessConceptLayer.NumeratorDenominatorPair;
+import rifServices.dataStorageLayer.common.AbstractSQLManager;
 import rifServices.dataStorageLayer.common.AgeGenderYearManager;
 import rifServices.dataStorageLayer.common.CovariateManager;
 import rifServices.dataStorageLayer.common.RIFContextManager;
@@ -26,8 +27,7 @@ import rifServices.system.RIFServiceError;
 import rifServices.system.RIFServiceMessages;
 import rifServices.system.RIFServiceStartupOptions;
 
-final class MSSQLInvestigationManager
-	extends MSSQLAbstractSQLManager {
+final class MSSQLInvestigationManager extends AbstractSQLManager {
 
 	private RIFContextManager rifContextManager;
 	private AgeGenderYearManager ageGenderYearManager;
