+++ rifServices/src/main/java/rifServices/dataStorageLayer/ms/MSSQLRIFContextManager.java
@@ -21,12 +21,13 @@ import rifServices.businessConceptLayer.GeoLevelView;
 import rifServices.businessConceptLayer.Geography;
 import rifServices.businessConceptLayer.HealthTheme;
 import rifServices.businessConceptLayer.NumeratorDenominatorPair;
+import rifServices.dataStorageLayer.common.AbstractSQLManager;
 import rifServices.dataStorageLayer.common.RIFContextManager;
 import rifServices.system.RIFServiceError;
 import rifServices.system.RIFServiceMessages;
 import rifServices.system.RIFServiceStartupOptions;
 
-final class MSSQLRIFContextManager extends MSSQLAbstractSQLManager implements RIFContextManager {
+final class MSSQLRIFContextManager extends AbstractSQLManager implements RIFContextManager {
 
 	/**
 	 * Instantiates a new SQLRIF context manager.
