+++ rifServices/src/main/java/rifServices/dataStorageLayer/ms/MSSQLRIFSubmissionManager.java
@@ -22,6 +22,7 @@ import rifServices.businessConceptLayer.NumeratorDenominatorPair;
 import rifServices.businessConceptLayer.RIFStudySubmission;
 import rifServices.businessConceptLayer.Sex;
 import rifServices.businessConceptLayer.YearRange;
+import rifServices.dataStorageLayer.common.AbstractSQLManager;
 import rifServices.dataStorageLayer.common.SampleTestObjectGenerator;
 import rifServices.dataStorageLayer.common.StudyStateManager;
 import rifServices.dataStorageLayer.common.SubmissionManager;
@@ -29,7 +30,7 @@ import rifServices.system.RIFServiceError;
 import rifServices.system.RIFServiceMessages;
 import rifServices.system.RIFServiceStartupOptions;
 
-public class MSSQLRIFSubmissionManager extends MSSQLAbstractSQLManager
+public class MSSQLRIFSubmissionManager extends AbstractSQLManager
 		implements SubmissionManager {
 
 	private StudyStateManager studyStateManager;
