+++ rifServices/src/main/java/rifServices/dataStorageLayer/ms/MSSQLResultsQueryManager.java
@@ -15,13 +15,13 @@ import rifGenericLibrary.system.RIFServiceException;
 import rifGenericLibrary.util.RIFLogger;
 import rifServices.businessConceptLayer.GeoLevelSelect;
 import rifServices.businessConceptLayer.Geography;
+import rifServices.dataStorageLayer.common.AbstractSQLManager;
 import rifServices.dataStorageLayer.common.ResultsQueryManager;
 import rifServices.system.RIFServiceError;
 import rifServices.system.RIFServiceMessages;
 import rifServices.system.RIFServiceStartupOptions;
 
-final class MSSQLResultsQueryManager extends MSSQLAbstractSQLManager
-		implements ResultsQueryManager {
+final class MSSQLResultsQueryManager extends AbstractSQLManager implements ResultsQueryManager {
 
 	RIFLogger rifLogger = RIFLogger.getLogger();
