+++ rifServices/src/main/java/rifServices/dataStorageLayer/ms/MSSQLStudyExtractManager.java
@@ -10,6 +10,7 @@ import rifGenericLibrary.businessConceptLayer.User;
 import rifGenericLibrary.dataStorageLayer.DatabaseType;
 import rifGenericLibrary.system.RIFServiceException;
 import rifServices.businessConceptLayer.RIFStudySubmission;
+import rifServices.dataStorageLayer.common.AbstractSQLManager;
 import rifServices.dataStorageLayer.common.GetStudyJSON;
 import rifServices.dataStorageLayer.common.RifZipFile;
 import rifServices.dataStorageLayer.common.StudyExtractManager;
@@ -17,8 +18,7 @@ import rifServices.system.RIFServiceError;
 import rifServices.system.RIFServiceMessages;
 import rifServices.system.RIFServiceStartupOptions;
 
-public class MSSQLStudyExtractManager extends MSSQLAbstractSQLManager
-		implements StudyExtractManager {
+public class MSSQLStudyExtractManager extends AbstractSQLManager implements StudyExtractManager {
 
 	private static String EXTRACT_DIRECTORY;
 	private static String TAXONOMY_SERVICES_SERVER;
