+++ rifServices/src/main/java/rifServices/dataStorageLayer/pg/PGSQLCreateStudySubmissionStep.java
@@ -31,6 +31,7 @@ import rifServices.businessConceptLayer.Project;
 import rifServices.businessConceptLayer.RIFStudySubmission;
 import rifServices.businessConceptLayer.Sex;
 import rifServices.businessConceptLayer.YearRange;
+import rifServices.dataStorageLayer.common.AbstractSQLManager;
 import rifServices.dataStorageLayer.common.DiseaseMappingStudyManager;
 import rifServices.dataStorageLayer.common.MapDataManager;
 import rifServices.system.RIFServiceError;
@@ -100,7 +101,7 @@ import rifServices.system.RIFServiceStartupOptions;
  *
  */
 
-final class PGSQLCreateStudySubmissionStep extends PGSQLAbstractSQLManager {
+final class PGSQLCreateStudySubmissionStep extends AbstractSQLManager {
 
 	private static final RIFLogger rifLogger = RIFLogger.getLogger();
 	private static String lineSeparator = System.getProperty("line.separator");
