+++ rifServices/src/main/java/rifServices/dataStorageLayer/pg/PGSQLDiseaseMappingStudyManager.java
@@ -24,6 +24,7 @@ import rifServices.businessConceptLayer.GeoLevelView;
 import rifServices.businessConceptLayer.Geography;
 import rifServices.businessConceptLayer.Investigation;
 import rifServices.businessConceptLayer.Project;
+import rifServices.dataStorageLayer.common.AbstractSQLManager;
 import rifServices.dataStorageLayer.common.DiseaseMappingStudyManager;
 import rifServices.dataStorageLayer.common.InvestigationManager;
 import rifServices.dataStorageLayer.common.RIFContextManager;
@@ -31,7 +32,7 @@ import rifServices.system.RIFServiceError;
 import rifServices.system.RIFServiceMessages;
 import rifServices.system.RIFServiceStartupOptions;
 
-final class PGSQLDiseaseMappingStudyManager extends PGSQLAbstractSQLManager
+final class PGSQLDiseaseMappingStudyManager extends AbstractSQLManager
 		implements DiseaseMappingStudyManager {
 	
 	private Messages GENERIC_MESSAGES = Messages.genericMessages();
