+++ rifServices/src/main/java/rifServices/dataStorageLayer/pg/PGSQLStudyExtractManager.java
@@ -10,6 +10,7 @@ import rifGenericLibrary.businessConceptLayer.User;
 import rifGenericLibrary.dataStorageLayer.DatabaseType;
 import rifGenericLibrary.system.RIFServiceException;
 import rifServices.businessConceptLayer.RIFStudySubmission;
+import rifServices.dataStorageLayer.common.AbstractSQLManager;
 import rifServices.dataStorageLayer.common.GetStudyJSON;
 import rifServices.dataStorageLayer.common.RifZipFile;
 import rifServices.dataStorageLayer.common.StudyExtractManager;
@@ -19,8 +20,7 @@ import rifServices.system.RIFServiceStartupOptions;
 
 //import rifGenericLibrary.dataStorageLayer.common.SQLFunctionCallerQueryFormatter;
 
-public class PGSQLStudyExtractManager extends PGSQLAbstractSQLManager
-		implements StudyExtractManager {
+public class PGSQLStudyExtractManager extends AbstractSQLManager implements StudyExtractManager {
 
 	private static String EXTRACT_DIRECTORY;
 	private static String TAXONOMY_SERVICES_SERVER;
